backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
openai==1.58.1
chromadb==0.5.23
sentence-transformers==3.3.1
newspaper3k==0.2.8
beautifulsoup4==4.12.3
gtts==2.5.4
moviepy==1.0.3
Pillow==11.1.0
deep-translator==1.11.4
python-dotenv==1.0.1
sse-starlette==2.2.1
lxml==5.3.0
lxml_html_clean==0.4.1
langchain==0.3.14
langchain-text-splitters==0.3.4
groq==0.25.0
playwright>=1.40.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
